Move kolibri.logger to kolibri.core.logger
