+++ kolibri/core/logger/__init__.py
@@ -1,0 +1,1 @@
+default_app_config = 'kolibri.core.logger.apps.KolibriLoggerConfig'
