+++ kolibri/core/logger/utils/user_data.py
@@ -18,10 +18,10 @@ from kolibri.core.auth.models import Facility
 from kolibri.core.auth.models import FacilityUser
 from kolibri.core.content.models import ContentNode
 from kolibri.core.lessons.models import Lesson
-from kolibri.logger.models import AttemptLog
-from kolibri.logger.models import ContentSessionLog
-from kolibri.logger.models import ContentSummaryLog
-from kolibri.logger.models import MasteryLog
+from kolibri.core.logger.models import AttemptLog
+from kolibri.core.logger.models import ContentSessionLog
+from kolibri.core.logger.models import ContentSummaryLog
+from kolibri.core.logger.models import MasteryLog
 
 logger = logging.getLogger(__name__)
