+++ kolibri/deployment/default/urls.py
@@ -36,7 +36,7 @@ urlpatterns = [
     url(r'', include('kolibri.core.content.urls')),
     url(r'^api/', include('kolibri.core.auth.api_urls')),
     url(r'^api/', include('kolibri.core.content.api_urls')),
-    url(r'^api/', include('kolibri.logger.api_urls')),
+    url(r'^api/', include('kolibri.core.logger.api_urls')),
     url(r'^api/', include('kolibri.tasks.api_urls')),
     url(r'^api/', include('kolibri.core.exams.api_urls')),
     url(r'^api/', include('kolibri.core.device.api_urls')),
