+++ kolibri/plugins/learn/serializers.py
@@ -7,8 +7,8 @@ from rest_framework.serializers import SerializerMethodField
 from kolibri.core.auth.models import Classroom
 from kolibri.core.exams.models import Exam
 from kolibri.core.lessons.models import Lesson
-from kolibri.logger.models import ContentSummaryLog
-from kolibri.logger.models import ExamLog
+from kolibri.core.logger.models import ContentSummaryLog
+from kolibri.core.logger.models import ExamLog
 
 
 class ExamProgressSerializer(ModelSerializer):
